requirements.txt
fastapi
uvicorn
pymongo
google-generativeai
python-multipart
